Adds in the ActiveX control for O3D. Review URL: http://codereview.chromium.org/160078
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@21492 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 8c2b005ffdc22c871638d0cfd356441741247bd5

+++ o3d/plugin/npapi_host_control/win/host_control.cc
@@ -515,7 +515,7 @@ STDMETHODIMP CHostControl::Load(IPropertyBag* property_bag,
   if (property_bag2) {
     ULONG property_count;
     if (SUCCEEDED(property_bag2->CountProperties(&property_count))) {
-      for (int x = 0; x < property_count; ++x) {
+      for (ULONG x = 0; x < property_count; ++x) {
         PROPBAG2 property = {0};
         ULONG properties_read = 0;
         if (SUCCEEDED(property_bag2->GetPropertyInfo(x, 1, &property,
